+++ nuxeo/client.py
@@ -121,9 +121,7 @@ class NuxeoClient(object):
         self._session.stream = True
         self.client_kwargs = kwargs
 
-        self.ssl_verify_needed = True
-        if "verify" in kwargs:
-            self.ssl_verify_needed = kwargs["verify"]
+        self.ssl_verify_needed = kwargs.get("verify", True)
 
         atexit.register(self.on_exit)
 
@@ -289,13 +287,14 @@ class NuxeoClient(object):
         )
 
         exc = None
-        ssl_verify_needed = self.ssl_verify_needed
 
-        if ssl_verify_needed:
-            ssl_verify_needed = ssl_verify
-        if ssl_verify_needed and "verify" in kwargs:
-            ssl_verify_needed = kwargs["verify"]
-        kwargs.pop("verify")
+        if not self.ssl_verify_needed:
+            ssl_verify = False
+
+        if "verify" in kwargs:
+            if not kwargs["verify"]:
+                ssl_verify = False
+            kwargs.pop("verify")
 
         try:
             resp = self._session.request(
@@ -304,7 +303,7 @@ class NuxeoClient(object):
                 headers=headers,
                 auth=auth,
                 data=data,
-                verify=ssl_verify_needed,
+                verify=ssl_verify,
                 **kwargs,
             )
             resp.raise_for_status()